requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
django-filter

